Move ReviewEventListener side-effect logic into ReviewService.applyReviewSideEffects()
Deleted ReviewEventListener. Post-review side effects (submission status update,
processing status update, evidence-to-mark linking) now run inline in
ReviewService.performReview() within the same transaction. Eliminates
unnecessary same-module event indirection.
